% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetted.R
\name{byContinent}
\alias{byContinent}
\title{Fetch countries by continent}
\usage{
byContinent(
  name = c("asia", "europe", "africa", "north america", "south america", "oceania"),
  full.list = TRUE
)
}
\arguments{
\item{name}{name of continent}

\item{full.list}{whether to return only name of country or full list}
}
\value{
country data list matching continent
}
\description{
Obtain country data matching language first one or more letters
}
\details{
Note that choices for names of continent includes 'asia','europe','africa','north america','south america','oceania'
}
\examples{
# task 1: get only names of countries that contains with "africa" or "AFrica"
# note that the search in case-insensitive
byContinent("africa", full.list = FALSE)

# task 2: get only names of countries that contains with "ASIA" or "asia"
byContinent("asia", full.list = FALSE)

# task 3: repeat task 2, but return full list for each country
byContinent("europe", full.list = TRUE)

}
