% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply.R
\docType{data}
\name{capitalOf}
\alias{capitalOf}
\title{Fetch the latest capital of a country}
\format{
An object of class \code{list} of length 193.
}
\usage{
capitalOf
}
\value{
a list containing all countries and their corresponding capital
}
\description{
With specified country name or names, get the associated capital
}
\examples{
# view the searchable countries, return first 6
head(names(capitalOf))

#task 0: check if the capital of japan is included
#should be all in lower case
grep("japan",names(capitalOf), value = TRUE)

#task 1: check the capital of nigeria
capitalOf$nigeria

#task 2: check the capital of united states
capitalOf$`united states`


#task 3: check capital of multiple countries
capitalOf[c("slovenia","romania","malaysia")]

#task 4: what if the capital is not available
capitalOf[c("randomcountry","mexico","luxembourg")]

}
\keyword{datasets}
