% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply.R
\docType{data}
\name{continentOf}
\alias{continentOf}
\title{Fetch the continent of a country}
\format{
An object of class \code{list} of length 193.
}
\usage{
continentOf
}
\value{
a list containing all countries and corresponding continents
}
\description{
Get the continent that a particular country belongs to
}
\examples{
# view the searchable countries
names(continentOf)

#task 1: view the continent of algeria
continentOf$algeria

#task 2: view the continent of nigeria
continentOf$nigeria

#task 3: view the continent of multiple countries
continentOf[c("niger","china","colombia")]

#task 4: if the continent is not available
continentOf[c("niger","china","randomtest")]

}
\keyword{datasets}
