% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_capital}
\alias{country_capital}
\title{Dataset of countries and their capitals}
\format{
A data frame with the capitals that countries belong to
\describe{
   \item{ID}{country identifiers}
   \item{capital}{capitals of various country}
   }
}
\source{
{Internal data consisting of countries and their capital}
}
\usage{
data(country_capital)
}
\value{
corresponding capitals of countries
}
\description{
Dataset containing country IDs and their capitals
}
\examples{
# load the population dataset
data(country_capital)
data(country_names)
finaldb <- cbind(country_names,country_capital)

# view content of the country and population
head(finaldb)


}
\references{
\url{https://www.wikipedia.org}
}
\seealso{
Data \code{\link{country_names}} for linkage of IDs with capital
}
\keyword{datasets}
