% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_continent}
\alias{country_continent}
\title{Dataset of countries and their continents}
\format{
A data frame with the continents that countries belong to
\describe{
   \item{ID}{country identifiers}
   \item{continent}{continents of various country}
   }
}
\source{
{Internal data consisting of countries and their continent}
}
\usage{
data(country_continent)
}
\value{
corresponding continents of countries
}
\description{
Dataset containing country IDs and their continents
}
\examples{
# load the continent dataset
data(country_continent)
data(country_names)
finaldb <- cbind(country_names,country_continent)

# view content of the country and population
head(finaldb)


}
\references{
\url{https://www.wikipedia.org}
}
\seealso{
Data \code{\link{country_names}} for linkage of IDs with continent
}
\keyword{datasets}
