% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_language}
\alias{country_language}
\title{Dataset of country official languages}
\format{
A data frame with the languages of countries
\describe{
   \item{ID}{Identifier for the countries}
   \item{officiallanguage}{official languages of countries}
   }
}
\source{
{Internal data consisting of all country IDs and their official languages}
}
\usage{
data(country_language)
}
\value{
country ID and official languages
}
\description{
Data for all countries and their associated languages
}
\examples{
# load the country ID and languages
data(country_language)

# view content of the countries languages
head(country_language)

}
\references{
\url{wikipedia.com}
}
\seealso{
\code{\link{country_names}} for linkage of IDs with country names
}
\keyword{countries}
\keyword{country}
\keyword{language}
\keyword{languages,}
\keyword{of}
