% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2country-package.R
\docType{data}
\name{country_money}
\alias{country_money}
\title{Data for Country Money}
\format{
A data frame with all countries and columns for currency
\describe{
   \item{x}{country name}
   }
}
\source{
{Internal data for currencies belonging to every country}
}
\usage{
country_money
}
\description{
Data for all country currency names and symbols
}
\examples{
data(country_money)

}
\keyword{datasets}
