% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_names}
\alias{country_names}
\title{Dataset of country names}
\format{
A data frame with the names of all countries
\describe{
   \item{ID}{Identifier for the countries}
   \item{name}{name of the countries}
   }
}
\source{
{Internal data consisting of all country names}
}
\usage{
data(country_names)
}
\value{
data for names of country
}
\description{
Data for all country currency names and associated ID
}
\examples{
# load the country names
data(country_names)

# view content of the countries currency
head(country_names)

}
\references{
\url{wikipedia.com}
}
\keyword{country}
\keyword{names}
