% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_population}
\alias{country_population}
\title{Dataset of countries and their latest population}
\format{
A data frame with the population of countries
\describe{
   \item{ID}{country identifiers}
   \item{population2023}{country population as of 2023}
   }
}
\source{
{Internal data consisting of countries and their population}
}
\usage{
data(country_population)
}
\value{
corresponding population size of countries
}
\description{
Dataset containing country IDs and their population by year
}
\examples{
# load the population dataset
data(country_population)
data(country_names)
finaldb <- cbind(country_names,country_population)

# view content of the country and population
head(finaldb)


}
\references{
\url{https://www.wikipedia.org}
}
\seealso{
Data \code{\link{country_names}} for linkage of IDs with country population size
}
\keyword{datasets}
