% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetted.R
\name{countryEndsWith}
\alias{countryEndsWith}
\title{Fetch countries that ends with specified characters}
\usage{
countryEndsWith(char, full.list = TRUE)
}
\arguments{
\item{char}{character to search for}

\item{full.list}{whether to return only name of country or full list}
}
\value{
country data list matching a specified character
}
\description{
Subset and return countries given a specified characters to search
}
\examples{
# task 1: get only names of countries that end with "A" or "a"
# note that the search in case-insensitive
countryEndsWith("A", full.list = FALSE)

# task 2: get only names of countries that end with "No" or "no"
countryEndsWith("no", full.list = FALSE)

# task 3: repeat task 2, but return full list for each country
countryEndsWith("no")

}
\seealso{
[countryStartsWith()] for country search starting with specified characters, and [countryHas()] for countries that contain specified characters.
}
