% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetted.R
\name{countryHas}
\alias{countryHas}
\title{Fetch countries that contains with specified characters}
\usage{
countryHas(char, full.list = TRUE)
}
\arguments{
\item{char}{character to search for}

\item{full.list}{whether to return only name of country or full list}
}
\value{
country data list matching content from a specified character
}
\description{
Subset to obtain data for countries containing specified characters to search
}
\examples{
# task 1: get only names of countries that contains with "ER" or "er"
# note that the search in case-insensitive
countryHas("er", full.list = F)

# task 2: get only names of countries that contains with "LAND" or "lAND" or "land"
countryHas("lAND", full.list = F)

# task 3: repeat task 2, but return full list for each country
countryHas("many", full.list = T)

}
\seealso{
\code{\link{countryStartsWith}} for country search starting with specified characters, and [countryEndsWith()] for countries that end with a specified characters.
}
