% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetted.R
\name{countryStartsWith}
\alias{countryStartsWith}
\title{Fetch countries that start with specified characters}
\usage{
countryStartsWith(char, full.list = TRUE)
}
\arguments{
\item{char}{character to search for}

\item{full.list}{whether to return only name of country or full list}
}
\value{
country data list matching a specified character
}
\description{
Subset and return countries given a specified characters to search
}
\examples{
# task 1: get only names of countries that start with "A" or "a"
# note that the search in case-insensitive
countryStartsWith("A", full.list = F)

# task 2: get only names of countries that start with "No" or "no"
countryStartsWith("no", full.list = F)

# task 3: repeat task 2, but return full list for each country
countryStartsWith("no")

}
\seealso{
[countryEndsWith()] for country search ending in specified character, and [countryHas()] for countries that contain specified characters.
}
