% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply.R
\docType{data}
\name{populationOf}
\alias{populationOf}
\title{Fetch the latest population of a country}
\format{
An object of class \code{list} of length 193.
}
\usage{
populationOf
}
\value{
a list containing all countries and their corresponding population
}
\description{
With specified country name or names, get the associated population
}
\examples{
# view the searchable countries, return first 6
head(names(populationOf))

#task 0: check if the population of japan is included
#should be all in lower case
grep("japan",names(populationOf), value = TRUE)

#task 1: check the population of nigeria
populationOf$nigeria

#task 2: check the population of united states
populationOf$`united states`


#task 3: check population of multiple countries
populationOf[c("slovenia","romania","malaysia")]

#task 4: what if the population is not available
populationOf[c("randomcountry","mexico","luxembourg")]

}
\keyword{datasets}
