% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2country-package.R
\docType{package}
\name{quickcode-package}
\alias{quickcode}
\alias{quickcode-package}
\title{quickcode: A Compilation of Some Frequently Used R Functions}
\description{
The NOT functions and a simple compilation of various functions for easy usage. Shorthand code to save memory usage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://quickcode.obi.obianom.com}
  \item Report bugs at \url{https://github.com/oobianom/quickcode}
}

}
\author{
\strong{Maintainer}: Obinna Obianom \email{idonshayo@gmail.com}

}
\keyword{internal}
