% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2country-package.R
\docType{package}
\name{r2country-package}
\alias{r2country}
\alias{r2country-package}
\title{r2country: Country Data with Names, Capitals, Currencies, Populations, Time, Languages and so on}
\description{
Obtain information about countries around the globe. Information for names, states, languages, time, capitals, currency and many more. Information were obtained from using 'Wikipedia' \url{https://www.wikipedia.org}, 'TimeAndDate' \url{https://www.timeanddate.com} and 'CountryCode' \url{https://countrycode.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r2country.obi.obianom.com}
  \item Report bugs at \url{https://github.com/oobianom/r2country}
}

}
\author{
\strong{Maintainer}: Obinna Obianom \email{idonshayo@gmail.com}

}
\keyword{internal}
