% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply.R
\docType{data}
\name{timeOf}
\alias{timeOf}
\title{Fetch the current time in a specific city}
\format{
An object of class \code{list} of length 2226.
}
\usage{
timeOf
}
\value{
a list containing all cities and corresponding local time
}
\description{
Get the current time using specified city name
}
\examples{
# view the searchable cities, return first 10
names(timeOf)[1:10]

#task 0: check if the time in Delhi is present
grep("Delhi",names(timeOf), value = TRUE)

#task 1: check the time in delhi
timeOf$delhi

#task 2: check the time in Boston
timeOf$boston

#task 3: view the time in multiple countries
timeOf[c("ankara","dubai","ontario")]

#task 4: what if the city is not available
timeOf[c("randomcity","dili","jerusalem")]

}
\keyword{datasets}
