% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city_time}
\alias{city_time}
\title{Dataset of current time in Cities}
\format{
A data frame with the current time in various cities
\describe{
   \item{City}{Names of cities}
   \item{Timediff}{Time difference relative to New York time}
   }
}
\source{
{Internal data consisting of cities and their corresponding time}
}
\usage{
data(city_time)
}
\value{
cities and the current time relative to New York
}
\description{
Dataset containing time different in various cities with respect to USA/New York time
}
\examples{
# load the cities and languages dataset
data(city_time)

# view content of the city-language data
head(city_time)

}
\references{
\url{https://www.timeanddate.com}
}
\keyword{datasets}
