% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\docType{data}
\name{COUNTRIES}
\alias{COUNTRIES}
\title{Names of all countries in upper case}
\format{
An object of class \code{character} of length 193.
}
\usage{
COUNTRIES
}
\value{
Names of all countries in the globe in upper case
}
\description{
Fetch the names of all countries in upper case
}
\examples{
# list all the countries
COUNTRIES

# list only 5 countries
COUNTRIES[1:5]

# create a dataset with all countries of the world
data.frame(ID = 1, Names = COUNTRIES)

}
\keyword{datasets}
