% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_calling_code}
\alias{country_calling_code}
\title{Dataset of countries and their calling code}
\format{
A data frame with the calling code of countries
\describe{
   \item{ID}{country identifiers}
   \item{callingcode}{Calling code of countries}
   }
}
\source{
{Internal data consisting of countries and their calling code}
}
\usage{
data(country_calling_code)
}
\value{
calling code of countries
}
\description{
Dataset containing country IDs and their calling code
}
\examples{
# load the calling code dataset
data(country_calling_code)

# view content of the city-calling code data
head(country_calling_code)


}
\references{
\url{timeanddate.com}
}
\seealso{
Data \code{\link{country_names}} for linkage of IDs with country calling code
}
\keyword{calling}
\keyword{code}
\keyword{code,}
\keyword{country}
\keyword{country,}
