% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_money}
\alias{country_money}
\title{Dataset for Country Currencies}
\format{
A data frame with all countries and columns for currency
\describe{
   \item{ID}{Identifier for countries}
   \item{currency}{name of the currency}
   \item{symbol}{symbol of the currency}
   \item{isocode}{The ISO code of the currency}
   \item{fractionalunity}{The fractional unit of the currency}
   }
}
\source{
{Internal data for currencies belonging to every country}
}
\usage{
data(country_money)
}
\value{
data for currencies by country
}
\description{
Data for all country currency names and symbols
}
\examples{
data(country_money)

# view content of the countries currency
head(country_money)

}
\references{
\url{https://www.wikipedia.org}
}
\seealso{
\code{\link{country_names}} for linkage of IDs with country names
}
\keyword{datasets}
